editoast: move get_paginated_errors to infra impl
